+++ packages/react/src/atomic/Button/src/components/ButtonAnchor/index.ts
@@ -1,4 +1,0 @@
-import { ButtonAnchor } from "./ButtonAnchor";
-
-export { ButtonAnchor } from "./ButtonAnchor";
-export default ButtonAnchor;