+++ packages/react/src/atomic/Button/src/components/ButtonLink/buttonLink.types.ts
@@ -4,6 +4,6 @@ import { ButtonProps } from "../../button.types";
 type Extend = Omit<ButtonProps, "ButtonHTMLAttributes" | "disabled"> &
   AnchorHTMLAttributes<HTMLAnchorElement>;
 
-export interface ButtonAnchorProps extends Extend {
+export interface ButtonLinkProps extends Extend {
   href: string;
 }