+++ packages/react/src/atomic/Button/src/components/index.ts
@@ -1,2 +1,2 @@
 export * from "./ButtonSkeleton";
-export * from "./ButtonAnchor";
+export * from "./ButtonLink";